generators: extract RFBucket matching into separate file
... in view of providing several other distribution types
for the longitudinal plane. The generators.py file will grow
too large and contain two parts: generating macro-particles
and matching the distributions in longitudinal. It seems logic
to split here.
